requirements.txt
gradio>=4.0.0
rembg>=2.0.0
pillow>=9.0.0
numpy>=1.21.0
onnxruntime>=1.12.0
spaces>=0.19.4
